Updated premake files, now release mode is actually release mode, assimp also updated, JobSystem(WIP). Also added resource access mutex(for graphics, due to vulkan doesn't allow use pools from multiplethreads at once.)

+++ Gauntlet/Source/Gauntlet/Renderer/Camera/PerspectiveCamera.h
@@ -99,8 +99,8 @@ class PerspectiveCamera : public Camera
     {
         if (m_FirstMouse)
         {
-            m_LastMouseX = InEvent.GetOffsetX();
-            m_LastMouseY = InEvent.GetOffsetY();
+            m_LastMouseX = (float)InEvent.GetOffsetX();
+            m_LastMouseY = (float)InEvent.GetOffsetY();
             m_FirstMouse = false;
         }
